+++ .gitignore
@@ -1,2 +1,4 @@
 node_modules
 ex.js
+index.js
+mutations.js
